update logging functionality
