+++ client.go
@@ -3,6 +3,7 @@ package mcp_golang
 import (
 	"context"
 	"encoding/json"
+
 	"github.com/metoro-io/mcp-golang/internal/protocol"
 	"github.com/metoro-io/mcp-golang/transport"
 	"github.com/pkg/errors"
